requirements.txt
autopep8==1.4.4
cycler==0.10.0
decorator==4.4.1
graphviz==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.2
networkx==2.4
numpy==1.17.4
Platypus-Opt==1.0.3
pycodestyle==2.5.0
pydot==1.4.1
pyparsing==2.4.5
python-dateutil==2.8.1
scipy==1.3.3
six==1.13.0
